MOBILE ENTITY WORKSSS pretty nice lol. The follow player and linear MovementTypes work, I just haven't tested avoid player yet but i think it's probably just follow with negative speed.
I simplified the getRect() and getEllipse() functions into one, getShape(), inherited by all GameComponent subclasses. This is so much better because instead of having both available and needing to denote which one to use with a variable for each GameComponent subclass (kinda cringe) and having to check a bunch of conditionals and stuff because ellipses and rects aren't compatible or something, I just put in getShape() and it returns the right shape type (rect/ellipse) based on the overridden return type of getShape() in the subclass.

EZ.